move job tests to CloneImagesOrVideosTest
